[Fix #3394] Separate Array & Hash Literal Comma configuration
