+++ lib/rubocop.rb
@@ -496,7 +496,8 @@ require_relative 'rubocop/cop/style/symbol_proc'
 require_relative 'rubocop/cop/style/ternary_parentheses'
 require_relative 'rubocop/cop/style/trailing_body_on_method_definition'
 require_relative 'rubocop/cop/style/trailing_comma_in_arguments'
-require_relative 'rubocop/cop/style/trailing_comma_in_literal'
+require_relative 'rubocop/cop/style/trailing_comma_in_array_literal'
+require_relative 'rubocop/cop/style/trailing_comma_in_hash_literal'
 require_relative 'rubocop/cop/style/trailing_method_end_statement'
 require_relative 'rubocop/cop/style/trailing_underscore_variable'
 require_relative 'rubocop/cop/style/trivial_accessors'
