+++ manual/cops.md
@@ -500,7 +500,8 @@ In the following section you find all available cops:
 * [Style/TernaryParentheses](cops_style.md#styleternaryparentheses)
 * [Style/TrailingBodyOnMethodDefinition](cops_style.md#styletrailingbodyonmethoddefinition)
 * [Style/TrailingCommaInArguments](cops_style.md#styletrailingcommainarguments)
-* [Style/TrailingCommaInLiteral](cops_style.md#styletrailingcommainliteral)
+* [Style/TrailingCommaInArrayLiteral](cops_style.md#styletrailingcommainarrayliteral)
+* [Style/TrailingCommaInHashLiteral](cops_style.md#styletrailingcommainhashliteral)
 * [Style/TrailingMethodEndStatement](cops_style.md#styletrailingmethodendstatement)
 * [Style/TrailingUnderscoreVariable](cops_style.md#styletrailingunderscorevariable)
 * [Style/TrivialAccessors](cops_style.md#styletrivialaccessors)
